Restructured the configuration option handling.
Still needs some more works and especially documentation but so far it
appears to work quite alright.

Also added test cases for the SES3D 4.0 writer and adapted it to the new
configuration handling. The munich ses3d writer is currently disabled
but that it ok because it was only good for an now obsolete old version.